refactor(home): refactor HomeScreen and related widgets for improved user interface
