+++ lib/features/home/presentation/home.dart
@@ -1,14 +1,13 @@
-import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:gap/gap.dart';
+import 'package:go_router/go_router.dart';
 import 'package:ionicons/ionicons.dart';
 
-import '../../../utils/context_extensions.dart';
-
-import '../../common/app_bar_gone.dart';
-import '../../common/bottom_nav_bar/bottom_nav_bar.dart';
+import '../../../../utils/context_extensions.dart';
+import '../../../common/bottom_nav_bar/bottom_nav_bar.dart';
 import 'widgets/header.dart';
+import 'widgets/language_tile.dart';
 import 'widgets/social_tile_widget.dart';
 import 'widgets/theme_widget.dart';
 
@@ -18,7 +17,15 @@ class HomeScreen extends ConsumerWidget {
   @override
   Widget build(BuildContext context, WidgetRef ref) {
     return Scaffold(
-      appBar: const EmptyAppBar(),
+      appBar: AppBar(
+        leading: IconButton(
+          icon: Icon(Icons.logout_outlined,
+              color: context.colorScheme.primary),
+          onPressed: () {
+            context.pop();
+          }
+        ),
+      ),
       bottomNavigationBar: const BottomNavBar(),
       backgroundColor: context.colorScheme.surface,
       body: Column(
@@ -73,32 +80,6 @@ class HomeScreen extends ConsumerWidget {
             },
           )
         ],
-      ),
-    );
-  }
-}
-
-class LanguageTile extends StatelessWidget {
-  const LanguageTile({
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return SwitchListTile(
-      onChanged: (bool newValue) {
-        /// Example: Change locale
-        /// The initial locale is automatically determined by the library.
-        /// Changing the locale like this will persist the selected locale.
-        context.setLocale(newValue ? const Locale('tr') : const Locale('en'));
-      },
-      shape: const RoundedRectangleBorder(
-          borderRadius: BorderRadius.all(Radius.circular(12))),
-      value: context.locale == const Locale('tr'),
-      title: Text(
-        tr('toggle_language'),
-        style:
-            Theme.of(context).textTheme.titleMedium!.apply(fontWeightDelta: 2),
       ),
     );
   }
